refactor(infra): move db/types and db/mappers to infra root and update imports
Refactored the folder structure under `@/infra` by moving `db/types` and `db/mappers` to the root of the `infra` directory.
Also updated all related import paths across the codebase to reflect the new locations.
This change improves maintainability and aligns better with the domain-infra separation.
