+++ adapters/cart/prisma/map-prisma-cart-to-domain-cart.ts
@@ -1,5 +1,5 @@
 import { CartEntity, newCartEntity } from "@/domain/entities/cart.entity";
-import { CartDatabase } from "@/infra/db/types/cart";
+import { CartDatabase } from "@/infra/types/cart";
 import { mapPrismaCartItemToDomainCartItem } from "@/adapters/cart/prisma/map-prisma-cart-item-to-domain-cart-item";
 import { newCurrencyEntity } from "@/domain/entities/currency.entity";
 import { newDateEntity } from "@/domain/entities/date.entity";
