+++ infra/services/cart/create-prisma-cart.service.ts
@@ -1,6 +1,6 @@
-import { cartPrismaToCartDatabaseMapper } from "@/infra/db/mappers/cart/cart-prisma-to-cart-database.mapper.ts";
+import { cartPrismaToCartDatabaseMapper } from "@/infra/mappers/cart/cart-prisma-to-cart-database.mapper.ts";
 import { prisma } from "@/infra/db/prisma";
-import { CartDatabase, CartDatabaseInsert } from "@/infra/db/types/cart";
+import { CartDatabase, CartDatabaseInsert } from "@/infra/types/cart";
 
 export async function createPrismaCartService(
   cart: CartDatabaseInsert
