+++ infra/services/product/get-prisma-latest-products.service.ts
@@ -3,7 +3,7 @@
 import { prisma } from "@/infra/db/prisma";
 import { prismaToJS } from "@/lib/utils";
 import { LATEST_PRODUCTS_LIMIT } from "@/lib/constants/product";
-import { ProductDatabase } from "@/infra/db/types/product";
+import { ProductDatabase } from "@/infra/types/product";
 
 export async function getPrismaLatestProductsService(): Promise<
   ProductDatabase[]
