+++ infra/services/product/get-prisma-product-by-id.service.ts
@@ -2,7 +2,7 @@
 
 import { prismaToJS } from "@/lib/utils";
 import { prisma } from "@/infra/db/prisma";
-import { ProductDatabase } from "@/infra/db/types/product";
+import { ProductDatabase } from "@/infra/types/product";
 
 export async function getPrismaProductByIdService(
   id: string
