+++ infra/types/product.ts
